more discretized
